sizeof

+++ hello_world/6-size.c
@@ -6,10 +6,10 @@
  */
 int main(void)
 {
-	printf("The size of a char: %lu byte(s)\n", sizeof(char));
-	printf("The size of an int: %lu byte(s)\n", sizeof(int));
-	printf("The size of a long int: %lu byte(s)\n", sizeof(long int));
-	printf("The size of a long long int: %lu byte(s)\n", sizeof(long long int));
-	printf("The size of a float: %lu byte(s)\n", sizeof(float));
+	printf("The size of a char: %lu byte(s)\n", (unsigned long)sizeof(char));
+	printf("The size of an int: %lu byte(s)\n", (unsigned long)sizeof(int));
+	printf("The size of a long int: %lu byte(s)\n", (unsigned long)sizeof(long int));
+	printf("The size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(long long int));
+	printf("The size of a float: %lu byte(s)\n", (unsigned long)sizeof(float));
 	return (0);
 }
